feat(grid): convert grid to xstyled
