+++ src/components/Grid/Box.jsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components'
-import { compose, space, layout, typography, color, flexboxes } from '@xstyled/system'
+import { compose, space, layout, typography, color, flexboxes, backgrounds } from '@xstyled/system'
 import { getSystemPropTypes } from '@xstyled/prop-types'
 
-const props = compose(space, layout, typography, color, flexboxes)
+const props = compose(space, layout, typography, color, flexboxes, backgrounds)
 
 const Box = styled.div`
   margin: 0;
