+++ src/components/Grid/Flex.jsx
@@ -1,4 +1,4 @@
-import styled from 'styled-components'
+import styled from '@xstyled/styled-components'
 
 import Box from './Box'
