+++ src/stories/Button.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks'
 
 import { Button } from '../components'
-import { ThemeProvider } from '../theme/xStyled'
+import { ThemeProvider } from '../theme'
 
 <Meta title="Button" component={Button} />
