+++ src/stories/Grid.stories.mdx
@@ -13,39 +13,39 @@ O componente Box é responsável por encapsular uma porção de conteúdo, afim 
 <Preview>
   <Story name="Flex | Box">
     <ThemeProvider>
-      <Flex flexWrap='wrap' backgroundColor='gray.main'>
+      <Flex flexWrap='wrap' backgroundColor='gray.400'>
         <Box width={1} p='2'>
           <Typography color='white'>Flex</Typography>
         </Box>
-        <Box width={{xs: 1, sm: 1 / 4}} p='2'>
-          <Typography color='white' backgroundColor='amber.main' textAlign='center' py='3'>Box 1 / 4</Typography>
+        <Box width={{xs: 1, sm: 1 / 4}} p='2' backgroundColor='amber.400' >
+          <Typography color='white' textAlign='center'>Box 1 / 4</Typography>
         </Box>
-        <Box width={{xs: 1, md: 1 / 4}} p='2'>
-          <Typography color='white' backgroundColor='amber.main' textAlign='center' py='3'>Box 1 / 4</Typography>
+        <Box width={{xs: 1, md: 1 / 4}} p='2' backgroundColor='amber.400' >
+          <Typography color='white' textAlign='center'>Box 1 / 4</Typography>
         </Box>
-        <Box width={{xs: 1, md: 1 / 4}} p='2'>
-          <Typography color='white' backgroundColor='amber.main' textAlign='center' py='3'>Box 1 / 4</Typography>
+        <Box width={{xs: 1, md: 1 / 4}} p='2' backgroundColor='amber.400' >
+          <Typography color='white' textAlign='center'>Box 1 / 4</Typography>
         </Box>
-        <Box width={{xs: 1, md: 1 / 4}} p='2'>
-          <Typography color='white' backgroundColor='amber.main' textAlign='center' py='3'>Box 1 / 4</Typography>
+        <Box width={{xs: 1, md: 1 / 4}} p='2' backgroundColor='amber.400' >
+          <Typography color='white' textAlign='center'>Box 1 / 4</Typography>
         </Box>
-        <Box width={{xs: 1, md: 1 / 3}} p='2'>
-          <Typography color='white' backgroundColor='red.main' textAlign='center' py='3'>Box 1 / 3</Typography>
+        <Box width={{xs: 1, md: 1 / 3}} p='2' backgroundColor='red.400' >
+          <Typography color='white' textAlign='center'>Box 1 / 3</Typography>
         </Box>
-        <Box width={{xs: 1, md: 1 / 3}} p='2'>
-          <Typography color='white' backgroundColor='red.main' textAlign='center' py='3'>Box 1 / 3</Typography>
+        <Box width={{xs: 1, md: 1 / 3}} p='2' backgroundColor='red.400' >
+          <Typography color='white' textAlign='center'>Box 1 / 3</Typography>
         </Box>
-        <Box width={{xs: 1, md: 1 / 3}} p='2'>
-          <Typography color='white' backgroundColor='red.main' textAlign='center' py='3'>Box 1 / 3</Typography>
+        <Box width={{xs: 1, md: 1 / 3}} p='2' backgroundColor='red.400' >
+          <Typography color='white' textAlign='center'>Box 1 / 3</Typography>
         </Box>
-        <Box width={{xs: 1, md: 1 / 2}} p='2'>
-          <Typography color='white' backgroundColor='blue.main' textAlign='center' py='3'>Box 1 / 2</Typography>
+        <Box width={{xs: 1, md: 1 / 2}} p='2' backgroundColor='blue.400' >
+          <Typography color='white' textAlign='center'>Box 1 / 2</Typography>
         </Box>
-        <Box width={{xs: 1, md: 1 / 2}} p='2'>
-          <Typography color='white' backgroundColor='blue.main' textAlign='center' py='3'>Box 1 / 2</Typography>
+        <Box width={{xs: 1, md: 1 / 2}} p='2' backgroundColor='blue.400' >
+          <Typography color='white' textAlign='center'>Box 1 / 2</Typography>
         </Box>
-        <Box width={1} p='2'>
-          <Typography color='white' backgroundColor='green.main' textAlign='center' py='3'>Box 1</Typography>
+        <Box width={1} p='2' backgroundColor='green.400' >
+          <Typography color='white' textAlign='center'>Box 1</Typography>
         </Box>
       </Flex>
     </ThemeProvider>
