+++ src/stories/Tag.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks'
 
 import { Tag } from '../components'
-import { ThemeProvider } from '../theme/xStyled'
+import { ThemeProvider } from '../theme'
 
 <Meta title="Tag" component={Tag} />
