+++ src/stories/Typography.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks'
 
 import { Typography, Heading, Subtitle, Paragraph, Caption } from '../components'
-import { ThemeProvider } from '../theme/xStyled'
+import { ThemeProvider } from '../theme'
 
 <Meta title="Typography" />
