--- src/theme/breakpoints.js
@@ -1,1 +1,0 @@
-export default ['768px', '1280px']
