--- src/theme/fontWeights.js
@@ -1,11 +1,0 @@
-const fontWeights = [400, 600]
-
-fontWeights.heading = fontWeights[1]
-fontWeights.subtitle = fontWeights[1]
-fontWeights.paragraph = fontWeights[0]
-fontWeights.inputTitle = fontWeights[1]
-fontWeights.button = fontWeights[1]
-fontWeights.caption = fontWeights[0]
-fontWeights.tag = fontWeights[1]
-
-export default fontWeights
