--- src/theme/xStyled/index.js
@@ -1,2 +1,0 @@
-export * from './Provider'
-export * from './Theme'
